+++ packages/ai/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @px-ui/ai
 
+## 8.0.0
+
+### Patch Changes
+
+- Updated dependencies [72accd0]
+  - @px-ui/core@4.7.0
+
+## 7.0.0
+
+### Patch Changes
+
+- Updated dependencies [8ac70a6]
+  - @px-ui/core@4.6.0
+
 ## 6.3.0
 
 ### Minor Changes