+++ packages/forms/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @px-ui/forms
 
+## 14.0.0
+
+### Minor Changes
+
+- 72accd0: Add sidebar primitive
+
+### Patch Changes
+
+- Updated dependencies [72accd0]
+  - @px-ui/core@4.7.0
+
+## 13.0.0
+
+### Minor Changes
+
+- 8ac70a6: Expose Tooltip portal props
+
+### Patch Changes
+
+- Updated dependencies [8ac70a6]
+  - @px-ui/core@4.6.0
+
 ## 12.0.2
 
 ### Patch Changes